requirements.txt
Flask==2.3.2  # Latest stable version of Flask
rembg==2.0.57
Pillow==9.1.1
gunicorn==20.1.0
